// source/AdventureWorksAIWorkspaceGUI/src/api/generated/model/reportMessageDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AdventureWorksAIWorkspaceAPI.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportMessageDtoRole } from './reportMessageDtoRole'

export interface ReportMessageDto {
  id: string
  role: ReportMessageDtoRole
  content: string
  sortOrder: number
  /** @nullable */
  relatedSqlQueryId?: string | null
  createdAt: string
}